+++ mariwheater/src/main/java/br/com/mariwheater/mariwheater/service/scheduling/NotificationScheduler.java
@@ -1,7 +1,6 @@
 package br.com.mariwheater.mariwheater.service.scheduling;
 
 import br.com.mariwheater.mariwheater.model.City;
-import br.com.mariwheater.mariwheater.model.Notifications;
 import br.com.mariwheater.mariwheater.service.account.AccountService;
 import br.com.mariwheater.mariwheater.service.city.CityService;
 import br.com.mariwheater.mariwheater.service.mail.MailService;
@@ -12,7 +11,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class NotificationService {
+public class NotificationScheduler {
 
     private final NotificationsService notificationsService;
 
@@ -22,7 +21,7 @@ public class NotificationService {
 
     private final CityService cityService;
 
-    public NotificationService(NotificationsService notificationsService, MailService mailService, AccountService accountService, CityService cityService) {
+    public NotificationScheduler(NotificationsService notificationsService, MailService mailService, AccountService accountService, CityService cityService) {
         this.notificationsService = notificationsService;
         this.mailService = mailService;
         this.accountService = accountService;
@@ -32,17 +31,9 @@ public class NotificationService {
     //@Scheduled(cron = "0 0 7,19 * * *") //De 12 em 12 horas
     @Scheduled(fixedRate = 65000) //1 Minuto
     //@Scheduled(fixedRate = 3600000) // -> One hour request
-    public void checkWeatherAndNotify() {
+    public void constructNotifications() {
         List<City> dangerousCities = cityService.getAllCitiesWithTemperatureIsDangerous();
         notificationsService.createAllNotifications(dangerousCities);
     }
 
-        //@Scheduled(cron = "0 20 7,19 * * *")//De 12 em 12 horas
-    @Scheduled(fixedRate = 120000)
-    public void sendNotificationMail() {
-        var notifitionsList = notificationsService.getAllNotifications();
-        for (Notifications n : notifitionsList) {
-            mailService.sendSimpleEmail("robsonmoura970@gmail.com", "Alerta de temperatura", n.getMessage());
-        }
-    }
 }
